Bug 1850611 - Create a JSON file source of truth for our design tokens
Introducing a JSON file that holds our design tokens source of truth in an agnostic format.
The plan is to generate desktop and mobile tokens from this file.
We will be exploring a build process as a next step, potentially relying on Style Dictionary.
This is the format we're playing with for now, but not set in stone.

Differential Revision: https://phabricator.services.mozilla.com/D189153

Depends on D194165