FIDEFE-4643 - Use value object in design-tokens.json
By using the composite design token value structure, we gain flexibility
in handling themed and non-themed tokens as well as other cases that
we are not aware of. For example, this composite value structure can be
used to handle prefers-contrast theming, forced-colors (HCM) theming,
etc.

We also add the ability to use "default" as the default value of tokens.
This handles the case where a token has the same value in light and dark
mode but has a different value in perfers-contrast or forced-colors
mode. Currently "default" is the highest priority property of the value
object. If you have "light", "dark" and "default" for a token's
value, then the "default" property will be in the generated token and
not the "light" or "dark" property values.